+++ CHANGELOG.md
@@ -11,6 +11,9 @@ All notable changes to this project will be documented in this file. This change
 - Update to Closure v20170910
 - Update build to allow alternate `xxd -i` implementation ([#549](https://github.com/mfikes/planck/issues/549))
 
+### Fixed
+- Single-dash command line args not passed to script ([#550](https://github.com/mfikes/planck/issues/550))
+
 ## [2.9.0] - 2017-11-14
 ### Changed
 - transit-cljs 0.8.243